+++ src/database/models/roles/capability.rs
@@ -1,9 +1,20 @@
+//! # Capability
+//!
+//! This module consist of the `Capability` struct, wich represent the database
+//! columns of the table `capabilities`
+//!
+//! There is also the `CapabilityMinima` struct, which can be used to insert
+//! data in the database.
+//!
+//! You can also find the list of the available capabilities
+
 use crate::database::schema::capabilities;
 use crate::database::schema::capabilities::dsl::capabilities as table;
 use crate::database::Data;
 use diesel::prelude::*;
 use diesel::MysqlConnection;
 
+/// The `Capability` struct is the usable type for what's in the database
 #[derive(Identifiable, Queryable, Associations, Serialize, Deserialize, Clone, Debug, PartialEq)]
 #[table_name = "capabilities"]
 pub struct Capability {
@@ -11,6 +22,8 @@ pub struct Capability {
     pub name: String,
 }
 
+/// The `CapabilityMinima` struct is only used while inserting a new capability
+/// in the database
 #[derive(Serialize, Deserialize, Debug, Insertable)]
 #[table_name = "capabilities"]
 pub struct CapabilityMinima {
@@ -18,12 +31,12 @@ pub struct CapabilityMinima {
 }
 
 impl Capability {
-    /* ------------------------------- STATIC ------------------------------ */
-
+    /// Constructor of `Capability` from a role id
     pub fn from_id(conn: &MysqlConnection, id: &u32) -> Option<Self> {
         table.find(id).first::<Self>(conn).ok()
     }
 
+    /// Constructor of `Capability` from a role name
     pub fn from_name(conn: &MysqlConnection, name: &str) -> Option<Self> {
         table
             .filter(capabilities::name.eq(&name))
@@ -31,18 +44,22 @@ impl Capability {
             .ok()
     }
 
+    /// Get all the capabilities in the database in an array of `Capability`
     pub fn all(conn: &MysqlConnection) -> Vec<Self> {
         table.load(conn).unwrap_or(vec![])
     }
 
-    /// Get the Capability record that fits the `minima` given.
+    /// Constructor of `Capability` that fits the `minima` given.
     pub fn select_minima(conn: &MysqlConnection, minima: &CapabilityMinima) -> Option<Self> {
-        table
-            .filter(capabilities::name.eq(&minima.name))
-            .first::<Self>(conn)
-            .ok()
+        Self::from_name(&conn, &minima.name)
     }
 
+    /// Add a new capability in database.
+    ///
+    /// This function should not be called outside a "seeder" or an update
+    /// mechanism for the application because the capabilities will be
+    /// hardcoded to check each and every feature's access, so it makes no
+    /// sense if one can add capability dynamically.
     pub fn insert_minima(conn: &MysqlConnection, minima: &CapabilityMinima) -> Data<Self> {
         if let Some(past) = Self::select_minima(conn, minima) {
             Data::Existing(past)
@@ -59,6 +76,7 @@ impl Capability {
     }
 }
 
+/// All the capabilities of the application
 pub const CAPABILITIES: [&str; 22] = [
     "user:delete",
     "user:update",