+++ src/routes/api/v1/mod.rs
@@ -1,3 +1,12 @@
+//! V1 controller
+//!
+//! The purpose of this module is to group every routes implemented on their
+//! first version.
+//!
+//! Every child module SHOULD implement his own `collect` function and those
+//! functions SHOULD be called on the `v1::collectt` function to group every
+//! routes in each and every child module.
+
 use crate::auth::Auth;
 use crate::http::responders::api::ApiResponse;
 use rocket::http::Status;
@@ -7,6 +16,8 @@ pub mod capabilities;
 pub mod role;
 pub mod roles;
 
+/// Collect every routes that this module needs to share with the application
+/// The name `collect` is a project convention
 pub fn collect() -> Vec<rocket::Route> {
     let auth_routes = auth::collect();
     let capabilities_routes = capabilities::collect();
@@ -22,6 +33,8 @@ pub fn collect() -> Vec<rocket::Route> {
     .concat()
 }
 
+/// Return the version of this module
+/// why though ?
 #[get("/api/v1", rank = 1)]
 pub fn version(_auth: Auth) -> ApiResponse {
     ApiResponse::new(