+++ src/routes/api/v1/role.rs
@@ -1,26 +1,43 @@
-use crate::database::models::roles::forms::RoleData;
+//! # Role controller
+//!
+//! Group the creation, update and deletion of roles
+
+use crate::auth::Auth;
 use crate::database::models::roles::{
     capability::Capability,
+    forms::RoleData,
     role::{Role, RoleMinima},
     role_capability::RelRoleCapability,
 };
 use crate::database::{DBConnection, Data};
 use crate::http::responders::api::ApiResponse;
 
-use crate::auth::Auth;
-
 use rocket::http::Status;
 
 use rocket_contrib::json::Json;
 
+/// Collect every routes that this module needs to share with the application
+/// The name `collect` is a project convention
 pub fn collect() -> Vec<rocket::Route> {
     routes!(create, update, delete)
 }
 
+/// Create a new role
+///
+/// This new role is the json serialization of the `RoleData` type
+/// The name cannot be empty
 #[post("/api/v1/role", format = "json", data = "<data>")]
 pub fn create(conn: DBConnection, _auth: Auth, data: Json<RoleData>) -> ApiResponse {
     // convert data into a usable type
     let role_data = data.into_inner();
+
+    // prevent the creation of empty role name
+    if &role_data.name == &"" {
+        return ApiResponse::error(
+            Status::UnprocessableEntity,
+            "The role name cannot be an empty string",
+        );
+    }
 
     // create a new roleminima object
     let new_role = RoleMinima {
@@ -48,6 +65,11 @@ pub fn create(conn: DBConnection, _auth: Auth, data: Json<RoleData>) -> ApiRespo
     ApiResponse::new(Status::Ok, json!({}))
 }
 
+/// Update an existing role
+///
+/// When updating, the caller MUST specify each and every details, because
+/// there is no smart mechanism implemented to perform a differencial update.
+/// The old values are removed and the new values are inserted.
 #[put("/api/v1/role/<role_id>", format = "json", data = "<data>", rank = 3)]
 pub fn update(conn: DBConnection, role_id: u32, data: Json<RoleData>) -> ApiResponse {
     let opt_role = Role::from_id(&conn, &role_id);
@@ -94,6 +116,10 @@ pub fn update(conn: DBConnection, role_id: u32, data: Json<RoleData>) -> ApiResp
     ApiResponse::new(Status::Ok, json!({}))
 }
 
+/// Delete an existing role
+///
+/// This will first remove every capabilities linked to this role
+/// then it will remove the role.
 #[delete("/api/v1/role/<role_id>")]
 pub fn delete(conn: DBConnection, role_id: u32) -> ApiResponse {
     let opt_role = Role::from_id(&conn, &role_id);
@@ -106,6 +132,7 @@ pub fn delete(conn: DBConnection, role_id: u32) -> ApiResponse {
         );
     }
     let role = opt_role.unwrap();
+
     // reset capabilities
     role.clear_capabilities(&conn);